Added Cars that work with a Behavior Tree
Added the class Car, it handles steering and rendering.
Added a BehaviorTree with the applicable behaviors to use with the cars.

Added a navigationCollider, so cars that would go too fast wouldn't go over the bounds of the grid; resulting in a crash.

Edited EGridGraph so certain things could be changed from the BT.

Changed some rendering options in BaseAgent.cpp that will allow the Car to be drawn on top of the grid.

Tried to mess around with other classes, didn't turn out to work, changes are kept in.
Added more gifs.